Police + NPC behavior
NPC runs away after seeing the player and then trigger something to generate police, while the police maps to the player until they reach the house and then chase the player down and eventually ending the gameplay scenario.
