+++ Assets/Materials/NPC placeholder.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 547fb57a02fe2f848b80596f497eb013
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
