+++ Assets/Prefabs/NPC.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7f27317e1aa3c9c46a3689d871d8a427
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
