+++ Assets/Scenes/Nhat's Playtest/Enemy Testing.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f658678f32a5a204ca241367ee5cc985
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
