+++ Assets/Scenes/Nhat's Playtest/Enemy Testing/NavMesh-NavMesh.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ecde2488dd953f243a99f59f9d528528
+guid: 754714068e4fa5548bcbde100f329a53
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 23800000
